Refactor import paths in test files for consistency
- Updated import statements in multiple test files to reflect the new directory structure.
- Changed imports from `gnssmultipath.readRinexObs` to `gnssmultipath.readers.readRinexObs`.
- Adjusted imports for `PickleHandler` from `gnssmultipath.PickleHandler` to `gnssmultipath.utils.PickleHandler`.
- Modified imports in `SkyPlotSummary` from `gnssmultipath.SkyPlotSummary` to `gnssmultipath.plot.SkyPlotSummary`.
- Ensured all relevant test files are aligned with the updated module organization.
