+++ src/gnssmultipath/plot/SkyPlotSummary.py
