+++ src/gnssmultipath/plot/__init__.py
@@ -1,0 +1,2 @@
+from .plotResults import plotResults, plotResults_dont_use_TEX, make_barplot, make_barplot_dont_use_TEX
+from .SkyPlotSummary import make_skyplot_summary
